scripts/requirements.txt
boto3>=1.34.0
pyyaml>=6.0
jsonschema>=4.21.0
